editing and updating of posters and their fields now works!
it requires two classes. PosterCreator goes through fields of posters
category, and check if the poster already has fields based on them, if
not builds them. PosterCleaner take care of 'empty' fields - remove
them not to make mess in the database - only filled field are saved in
the database.
WARNING: PosterCleaner does not yet work properly for all the models
(like FloatField) - needs to be fixed!!!!